lincheck.cl 2024-02-23 version
Current version including ZEROCOMBINE to create a master bias image and IMARITH to subtract this bias image from all images. 

I tested the script and columnSep.sh and they both work well "parked" in the directory ABOVE WHICH the images reside. For example, if the images 
are in the directory:

images/g23d320/

and the scripts are in directory 

images/

both lincheck.cl and columnSep.sh work, after I made some minor edits. 

THese are now ready to produce the OVERSCAN subtracted data for PTC and DTC analysis as presented in Chapters 5 and 11 of Janesick's "Photon Transfer DN -> Lambda" book.